# doc/api.yml
openapi: 3.0.3
info:
  title: HEI Admin API
  description: |
    Management of HEI students, promotions, tracks, groups, courses, teachers,
    exams, grades and graduates.

    This file is the single source of truth for the project data model.
    Invariants that cannot be expressed in OpenAPI (foreign keys, uniqueness,
    non-overlapping periods) are documented in each schema description and will
    be materialized by the Flyway migrations and the business services.
  version: 1.0.0
servers:
  - url: https://api.hei-admin.com
    description: Production
  - url: http://localhost:8080
    description: Local
security:
  - BearerAuth: []
...
paths:
...
  /students/{id}/deactivation:
    post:
      tags: [ Account ]
      summary: Disable the account of a student
      description: |
        Reversible, and immediate: the holder is refused on their next call even
        though the token they already hold has not expired. The profile, its
        grades and its history are all kept.
      operationId: deactivateStudentAccount
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
  /students/{id}/activation:
    post:
      tags: [ Account ]
      summary: Enable the account of a student again
      operationId: activateStudentAccount
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
  /teachers/{id}/deactivation:
    post:
      tags: [ Account ]
      summary: Disable the account of a teacher
      operationId: deactivateTeacherAccount
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
  /teachers/{id}/activation:
    post:
      tags: [ Account ]
      summary: Enable the account of a teacher again
      operationId: activateTeacherAccount
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "403":
          $ref: "#/components/responses/403"
        "404":
          $ref: "#/components/responses/404"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    "403":
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    "404":
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Role:
      type: string
      description: Role carried by the user account.
      enum: [ STUDENT, TEACHER, ADMIN ]
...
    Account:
      description: |
        The sign-in side of a person. An account is disabled, never deleted:
        grade_history references it as the author of every change.
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/Role"
        enabled:
          type: boolean
...
    Error:
      type: object
      properties:
        type:
          type: string
          example: BadRequestException
        message:
          type: string