# src/main/java/com/exam/hei/endpoint/rest/model/Account.java
package com.exam.hei.endpoint.rest.model;

import com.exam.hei.repository.model.Role;
import com.fasterxml.jackson.databind.PropertyNamingStrategies;
import com.fasterxml.jackson.databind.annotation.JsonNaming;
import java.util.UUID;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;
import lombok.ToString;

@Getter
@Setter
@Builder
@NoArgsConstructor
@AllArgsConstructor
@ToString
@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
public class Account {
  private UUID userId;
  private String email;
  private Role role;
  private boolean enabled;
}